+++ website/gatsby-config.js
@@ -18,12 +18,6 @@ module.exports = {
     repositoryUrl:
       'https://github.com/commercetools/merchant-center-application-kit',
     excludeFromSearchIndex: false,
-    globalNotification: {
-      active: true,
-      notificationType: 'info',
-      content:
-        'This is the new documentation of Custom Applications. You can still visit the [legacy documentation](https://docs.commercetools.com/custom-applications/legacy) during the [migration](https://docs.commercetools.com/custom-applications/migrating-from-project-level-custom-applications) from Project-level Custom Applications.',
-    },
   },
   plugins: [
     // Docs theme