+++ packages/create-mc-app/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @commercetools-frontend/create-mc-app
 
+## 21.15.0
+
 ## 21.14.3
 
 ## 21.14.2