+++ packages/l10n/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @commercetools-frontend/l10n
 
+## 21.15.0
+
+### Patch Changes
+
+- Updated dependencies [[`0e49a78f8`](https://github.com/commercetools/merchant-center-application-kit/commit/0e49a78f8e3b228a9b0bb3d90781aa8e940de4bc)]:
+  - @commercetools-frontend/sentry@21.15.0
+
 ## 21.14.3
 
 ### Patch Changes