+++ website-components-playground/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @commercetools-website/components-playground
 
+## 21.15.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-frontend/i18n@21.15.0
+  - @commercetools-frontend/application-components@21.15.0
+
 ## 21.14.3
 
 ### Patch Changes